fix bug, 这个计算结果不可以为负，需要定义为无符号整型 unsigned int index = bucket->mHash % bucketSize;

+++ como/runtime/util/hashmap.h
@@ -75,7 +75,7 @@ private:
             mNext = nullptr;
         }
 
-        int mHash;
+        unsigned int mHash;
         Key mKey;
         Val mValue;
         struct Bucket* mNext;
@@ -87,7 +87,7 @@ public:
     using HashMapWalker = void(*)(String&,Key,Val);
 
     HashMap(
-        /* [in] */ int size = 50)
+        /* [in] */ unsigned int size = 50)
         : mCount(0)
     {
         mBucketSize = get_next_prime(size);
@@ -99,7 +99,7 @@ public:
 
     ~HashMap()
     {
-        for (int i = 0;  i < mBucketSize;  i++) {
+        for (unsigned int i = 0;  i < mBucketSize;  i++) {
             if (mBuckets[i] != nullptr) {
                 Bucket* curr = mBuckets[i];
                 while (curr != nullptr) {
@@ -125,11 +125,11 @@ public:
             Rehash();
         }
 
-        int hash = HashKey(key);
+        unsigned int hash = HashKey(key);
         if (hash == -1)
             return -1;
 
-        int index = (unsigned int)hash % mBucketSize;
+        unsigned int index = hash % mBucketSize;
         if (mBuckets[index] == nullptr) {
             Bucket* b = new Bucket();
             if (nullptr == b)
@@ -172,14 +172,15 @@ public:
     {
         CompareFunc<Key> compareF;
 
-        int hash = HashKey(key);
+        unsigned int hash = HashKey(key);
         if (hash == -1)
             return false;
 
-        int index = (unsigned int)hash % mBucketSize;
+        unsigned int index = (unsigned int)hash % mBucketSize;
         Bucket* curr = mBuckets[index];
         while (curr != nullptr) {
-            if (curr->mHash == hash && !compareF(curr->mKey, key)) return true;
+            if (curr->mHash == hash && !compareF(curr->mKey, key))
+                return true;
             curr = curr->mNext;
         }
 
@@ -191,11 +192,11 @@ public:
     {
         CompareFunc<Key> compareF;
 
-        int hash = HashKey(key);
+        unsigned int hash = HashKey(key);
         if (hash == -1)
             return Val(0);
 
-        int index = (unsigned int)hash % mBucketSize;
+        unsigned int index = hash % mBucketSize;
         Bucket* curr = mBuckets[index];
         while (curr != nullptr) {
             if (curr->mHash == hash && !compareF(curr->mKey, key))
@@ -211,11 +212,11 @@ public:
     {
         CompareFunc<Key> compareF;
 
-        int hash = HashKey(key);
+        unsigned int hash = HashKey(key);
         if (hash == -1)
             return;
 
-        int index = (unsigned int)hash % mBucketSize;
+        unsigned int index = hash % mBucketSize;
         Bucket* curr = mBuckets[index];
         Bucket* prev = curr;
         while (curr != nullptr) {
@@ -239,7 +240,7 @@ public:
 
     void Clear()
     {
-        for (int i = 0; i < mBucketSize; i++) {
+        for (unsigned int i = 0; i < mBucketSize; i++) {
             if (mBuckets[i] != nullptr) {
                 Bucket* curr = mBuckets[i];
                 while (curr != nullptr) {
@@ -256,7 +257,7 @@ public:
     Array<Val> GetValues()
     {
         Long N = 0;
-        for (int i = 0; i < mBucketSize; i++) {
+        for (unsigned int i = 0; i < mBucketSize; i++) {
             if (mBuckets[i] != nullptr) {
                 Bucket* curr = mBuckets[i];
                 while (curr != nullptr) {
@@ -273,7 +274,7 @@ public:
 
         if (N > 0) {
             Long idx = 0;
-            for (int i = 0;  i < mBucketSize;  i++) {
+            for (unsigned int i = 0;  i < mBucketSize;  i++) {
                 if (mBuckets[i] != nullptr) {
                     Bucket* curr = mBuckets[i];
                     while (curr != nullptr) {
@@ -289,7 +290,7 @@ public:
 
     void GetValues(String& str, HashMapWalker walker)
     {
-        for (int i = 0;  i < mBucketSize;  i++) {
+        for (unsigned int i = 0;  i < mBucketSize;  i++) {
             if (mBuckets[i] != nullptr) {
                 Bucket* curr = mBuckets[i];
                 while (curr != nullptr) {
@@ -300,13 +301,13 @@ public:
         }
     }
 
-    int GetDataNumber()
+    unsigned int GetDataNumber()
     {
         return mCount;
     }
 
 private:
-    int HashKey(
+    unsigned int HashKey(
         /* [in] */ const Key& key)
     {
         HashFunc<Key> hashF;
@@ -318,7 +319,7 @@ private:
         if (mBucketSize == MAX_BUCKET_SIZE) {
             return;
         }
-        int oldBucketSize = mBucketSize;
+        unsigned int oldBucketSize = mBucketSize;
         mBucketSize = oldBucketSize * 2 + 1;
         if (mBucketSize > MAX_BUCKET_SIZE || mBucketSize < 0) {
             mBucketSize = MAX_BUCKET_SIZE;
@@ -327,7 +328,7 @@ private:
         if (newBuckets == nullptr) {
             return;
         }
-        for (int i = 0; i < oldBucketSize; i++) {
+        for (unsigned int i = 0; i < oldBucketSize; i++) {
             Bucket* curr = mBuckets[i];
             while (curr != nullptr) {
                 Bucket* next = curr->mNext;
@@ -343,10 +344,10 @@ private:
 
     void PutBucket(
         /* [in] */ Bucket** buckets,
-        /* [in] */ int bucketSize,
+        /* [in] */ unsigned int bucketSize,
         /* [in] */ Bucket* bucket)
     {
-        int index = bucket->mHash % bucketSize;
+        unsigned int index = bucket->mHash % bucketSize;
         if (buckets[index] == nullptr) {
             buckets[index] = bucket;
             return;
@@ -364,9 +365,9 @@ private:
 private:
     static constexpr float LOAD_FACTOR = 0.75;
     static constexpr int MAX_BUCKET_SIZE = INT32_MAX - 8;
-    int mThreshold;
-    int mCount;
-    int mBucketSize;
+    unsigned int mThreshold;
+    unsigned int mCount;
+    unsigned int mBucketSize;
     Bucket** mBuckets;
 };
